test(tooling): bind the hosted readiness budget to its restore contract
